fix(code_suggestions): Detect binary correctly
In
https://gitlab.com/gitlab-org/editor-extensions/gitlab-lsp/-/merge_requests/322,
the path to the binary changed from `"./out/node/main.js"` to
`"./out/node/main-bundle.js"`. This meant that the plugin would never
detect the language server was installed, even when it was (via
`GitLabCodeSuggestionsInstallLanguageServer`).

Now we parse the binary's location from `@gitlab-org/gitlab-lsp`'s
package.json instead of hard-coding its location.